requirements.txt
discord.py==1.7.3
python-dotenv==0.18.0
requests>= 2.18.4
APScheduler==3.7.0
ujson==4.0.2
dhooks==1.1.4

